Moving the trace logger that uses profile manager so it can be separated out from xdp_core
