+++ src/runtime_src/xdp/CMakeLists.txt
@@ -31,6 +31,7 @@ set(XRT_XDP_PROFILE_WRITER_DIR         "${CMAKE_CURRENT_SOURCE_DIR}/profile/writ
 set(XRT_XDP_PROFILE_HAL_WRITER_DIR     "${CMAKE_CURRENT_SOURCE_DIR}/profile/writer/hal")
 set(XRT_XDP_PROFILE_CORE_DIR           "${CMAKE_CURRENT_SOURCE_DIR}/profile/core")
 set(XRT_XDP_PROFILE_DEVICE_DIR         "${CMAKE_CURRENT_SOURCE_DIR}/profile/device")
+set(XRT_XDP_PROFILE_OCL_DEVICE_DIR     "${CMAKE_CURRENT_SOURCE_DIR}/profile/device/ocl_device_logger")
 set(XRT_XDP_PROFILE_HAL_DEVICE_DIR     "${CMAKE_CURRENT_SOURCE_DIR}/profile/device/hal_device")
 set(XRT_XDP_PROFILE_XRT_DEVICE_DIR     "${CMAKE_CURRENT_SOURCE_DIR}/profile/device/xrt_device")
 set(XRT_XDP_PROFILE_LOP_PLUGIN_DIR     "${CMAKE_CURRENT_SOURCE_DIR}/profile/plugin/lop")
@@ -83,6 +84,8 @@ file(GLOB XRT_XDP_PROFILE_FILES
   "${XRT_XDP_PROFILE_CORE_DIR}/*.cpp"
   "${XRT_XDP_PROFILE_DEVICE_DIR}/*.h"
   "${XRT_XDP_PROFILE_DEVICE_DIR}/*.cpp"
+  "${XRT_XDP_PROFILE_OCL_DEVICE_DIR}/*.h"
+  "${XRT_XDP_PROFILE_OCL_DEVICE_DIR}/*.cpp"
   "${XRT_XDP_PROFILE_XRT_DEVICE_DIR}/*.h"
   "${XRT_XDP_PROFILE_XRT_DEVICE_DIR}/*.cpp"
   "${XRT_XDP_PROFILE_PLUGIN_DIR}/*.h"
