+++ src/runtime_src/xdp/profile/plugin/ocl/ocl_profiler.cpp
@@ -29,7 +29,7 @@
 #include "xdp/profile/profile_config.h"
 #include "xdp/profile/core/rt_profile.h"
 #include "xdp/profile/device/xrt_device/xdp_xrt_device.h"
-#include "xdp/profile/device/profile_mngr_trace_logger.h"
+#include "xdp/profile/device/ocl_device_logger/profile_mngr_trace_logger.h"
 #include "xdp/profile/device/tracedefs.h"
 #include "xdp/profile/writer/json_profile.h"
 #include "xdp/profile/writer/csv_profile.h"
